+++ dacite/CHANGELOG.md
@@ -22,6 +22,7 @@
    - `core::RenderPassCreateInfo`
    - `core::RenderPassBeginInfo`
    - `khr_swapchain::SwapchainCreateInfoKhr`
+   - `khr_swapchain::PresentInfoKhr`
 
  - The builder `core::SpecializationInfoBuilder` has been removed. Entries to a
    `core::SpecializationInfo` can now be added conveniently with `core::SpecializationInfo::push`.